Add Ansible support for Arch setup
Previous installation/uninstallation of configurations for apps like
Neovim, i3, etc used a combination of manual operations with a package
manager and shell scripts, which resulted in a fragile and error-prone
setup process. Moving to Ansible playbooks provides an idempotent setup
process that can automate most of what was being done before. The old
configurations are still in place for now, but plans are to remove these
in the future.